Move dse/clusterizer -> ir/clusters/...
