+++ devito/dse/backends/advanced.py
@@ -2,9 +2,9 @@ from __future__ import absolute_import
 
 from collections import OrderedDict
 
+from devito.ir import clusterize
 from devito.dse.aliases import collect
 from devito.dse.backends import BasicRewriter, dse_pass
-from devito.dse.clusterizer import clusterize
 from devito.symbolics import Eq, estimate_cost, xreplace_constrained, iq_timeinvariant
 from devito.dse.manipulation import (common_subexprs_elimination, collect_nested,
                                      compact_temporaries)
