+++ devito/ir/__init__.py
@@ -1,1 +1,2 @@
 from devito.ir.dfg import *  # noqa
+from devito.ir.clusters import *  # noqa
