+++ devito/ir/clusters/__init__.py
@@ -1,0 +1,1 @@
+from devito.ir.clusters.algorithms import *  # noqa
